requirements.txt
bcrypt==4.0.0
cryptography==38.0.1
jsonschema==4.16.0
ops==1.5.2
ruamel.yaml==0.17.21
